Split client to multiple files, one per resource root type
